Delete unused BufferManager class (we might revive it later). Simplify some code using macros (getters/setters implementation). Disable temporary fog hack code. Implemented basic billboarding support. Editor now draws translate gizmo (3D manipulators).

+++ src/Engine/Cell.cpp
@@ -55,8 +55,8 @@ void Cell::updateGeometry( )
 	const float tileWidth = (float(CellSize) / TileDimensions);
 	const float tileHeight = (float(CellSize) / TileDimensions);
 
-	float dX = x * CellSize;
-	float dZ = y * CellSize;
+	double dX = x * CellSize;
+	double dZ = y * CellSize;
 	
 	float height;
